backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
sqlalchemy
pydantic
transformers
torch
sentencepiece
